requirements.txt
streamlit
huggingface_hub
python-docx